requirements.txt
torch==2.2.1
torchvision==0.17.1
numpy==1.26.4
Pillow==10.2.0
matplotlib==3.8.3
tqdm==4.66.2
scikit-learn==1.4.1.post1
opencv-python==4.9.0.80 